% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Y_CT}
\alias{Y_CT}
\title{Clonal tracking data from clinical trials}
\format{
A list containing the clonal tracking data for each clinical trial (WAS, \eqn{\beta 0 \beta E}{b0bE}, \eqn{\beta S \beta S}{bSbS}).
Each clonal tracking dataset is a 3-dimensional array whose dimensions identify
\describe{
\item{1}{time, in months}
\item{2}{cell types: T, B, NK, Macrophages(M) and Granulocytes(G)}
\item{3}{unique barcodes (clones)}
}
}
\source{
\url{https://github.com/BushmanLab/HSC_diversity}
}
\usage{
Y_CT
}
\description{
A dataset containing clonal tracking cell counts from three different clinical trials.
}
\keyword{datasets}
